Harden backtest data validation and cash accrual
- Reject non-finite and non-positive prices, NAV paths, forecast paths, portfolio state inputs, and scenario/cash returns
- Accrue cash and holding costs over actual calendar-day gaps using configured day-count conventions
- Treat optimization failures as typed failures so fallback periods preserve solver status
- Require exact optimal solver status and report non-converged CVaR cutting-plane runs explicitly
- Fail loudly when realised costs drive cash materially negative instead of allowing overdrafts
- Infer daily cadence as 252 periods per year and fix WalkForwardConfig inherited validation
- Add regression tests for validation errors, calendar-gap costs, cash accrual, optimizer fallback status, and cadence inference

+++ qraft/backtest/costs.py
@@ -52,7 +52,7 @@ class CostModel:
         self,
         weights: NDArray[np.floating],
         nav: float,
-        n_periods: int = 1,
+        n_periods: float = 1.0,
     ) -> float:
         if self.holding is None or nav <= 0.0:
             return 0.0
